+++ src/Messages.js
@@ -1,44 +1,46 @@
 export default {
 
+    array: 'The :field field must be of type Array.',
+
+    between: 'The :field field must be between :param0 and :param1.',
+
     boolean: 'The :field field must be of type Boolean.',
 
-    string: 'The :field field must be of type String.',
+    confirmed: 'The :field field must be confirmed.',
+
+    different: 'The :field field value must be different than the :param0 field value.',
+    
+    email: 'The :field field must be a valid email address.',
+
+    equal: 'The :field field must be equal to :param0.',
+
+    filled: 'The :field field must be present and not be empty.',
+
+    gt: 'The :field field must be greater than :param0.',
+
+    gte: 'The :field field must be greater than or equal to :param0.',
+
+    in: 'The :field field must be in the given array.',
 
     integer: 'The :field field must be of type Integer.',
 
-    object: 'The :field field must be of type Object.',
+    length: 'The :field field must be of length :param0.',
+
+    lt: 'The :field field must be less than :param0.',
+
+    lte: 'The :field field must be less than or equal to :param0.',
 
     max: 'The :field field must be less than :param0.',
 
     min: 'The :field field must be greater than :param0.',
 
-    required: 'The :field field is required.',
-
     null: 'The :field field must be null.',
 
-    length: 'The :field field must be of length :param0.',
+    numeric: 'The :field field must be a number',
 
-    array: 'The :field field must be of type Array.',
+    object: 'The :field field must be of type Object.',
 
-    equal: 'The :field field must be equal to :param0.',
-
-    gt: 'The :field field must be greater than :param0.',
-
-    lt: 'The :field field must be less than :param0.',
-
-    gte: 'The :field field must be greater than or equal to :param0.',
-
-    lte: 'The :field field must be less than or equal to :param0.',
-
-    between: 'The :field field must be between :param0 and :param1.',
-
-    in: 'The :field field must be in the given array.',
-
-    same: 'The :field field must be equal to :param0.',
-
-    different: 'The :field field value must be different than the :param0 field value.',
-
-    confirmed: 'The :field field must be confirmed.',
+    required: 'The :field field is required.',
 
     required_if: 'The :field field is required if the :param0 field equals :param1.',
 
@@ -47,4 +49,8 @@ export default {
     required_with: 'The :field field is required with at least one of the listed fields.',
 
     required_with_all: 'The :field field is required with all of the listed fields.',
+
+    same: 'The :field field must be equal to :param0.',
+
+    string: 'The :field field must be of type String.',
 }