+++ webpack.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-//
+
 module.exports = {
   mode: 'production',
   entry: './src/index.js',